Three 2D Gaussians Included
I have included a last example at the end of the notebok "00_Explore_2D_Gaussians_Functions_Fits.ipynb" which includes the fit of three 2D gaussians. This shows the obvious effect that the initial guess of the parameters have on the number of iterations.  I also included this last example as a separate file named "04_Three_2DGauss.py" under the Py_Files/00_Explore_2D_Gaussians_Functions_Fits folder. The output of such example is a plot of the fit-result to "true" gaussians subtraction.